% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_gbif_backbone_taxonomy.R
\name{add_gbif_backbone_taxonomy}
\alias{add_gbif_backbone_taxonomy}
\title{add_gbif_backbone_taxonomy}
\usage{
add_gbif_backbone_taxonomy(dataframe, speciescolumn)
}
\arguments{
\item{dataframe}{a dataframe containing a species binomial name}

\item{speciescolumn}{the column where the species binomial name}
}
\value{
a dataframe with added species gbif taxonomy
}
\description{
add_gbif_backbone_taxonomy
}
