% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_merge_clean_data_files.R
\name{format_merged}
\alias{format_merged}
\title{Add water percent and transform tonnage to tons of C, N, P}
\usage{
format_merged(tibble_merged)
}
\value{
unique tibble to work on, with tonnage and equivalent
export in C, N,P
}
\description{
Add water percent and transform tonnage to tons of C, N, P
}
