% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_plots.R
\name{graphton_nb_countries}
\alias{graphton_nb_countries}
\title{Create graph f(total_tonnage) = nb_countries}
\usage{
graphton_nb_countries(data)
}
\value{
ggplot object
}
\description{
Create graph f(total_tonnage) = nb_countries
}
