% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_wrangle_ghs_data.R
\name{wrangling_ghs_data}
\alias{wrangling_ghs_data}
\title{wrangling_ghs_data}
\usage{
wrangling_ghs_data(pat)
}
\value{
a tibble containing species ID, taxonomy and chemical composition
}
\description{
wrangling_ghs_data
}
\examples{


}
